move api_transforms into the model
