+++ iml-wasm-components/iml-action-dropdown/src/fetch_actions.rs
@@ -2,7 +2,10 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-use crate::{api_transforms::composite_ids_to_query_string, Msg, RecordMap};
+use crate::{
+    model::{composite_ids_to_query_string, RecordMap},
+    update::Msg,
+};
 use futures::Future;
 use seed::Request;
